+++ packages/xchain-mayachain-query/__mocks__/mayanode-api.ts
@@ -21,7 +21,20 @@ export default {
         return [200, resp]
       }
 
+      if (from_asset === 'ETH.ETH' && to_asset === 'BTC.BTC') {
+        const resp = require(`./responses/mayanode/QuoteSSwapEthBtc.json`)
+        return [200, resp]
+      }
+
       return [200, {}]
+    })
+    mock.onGet(/\/mayachain\/mimir/).reply(function () {
+      const resp = require(`./responses/mayanode/mimir.json`)
+      return [200, resp]
+    })
+    mock.onGet(/\/mayachain\/lastblock/).reply(function () {
+      const resp = require(`./responses/mayanode/latestBlock.json`)
+      return [200, resp]
     })
   },
 }