+++ examples/liquidity/CHANGELOG.md
@@ -1,5 +1,30 @@
 # xchainjs-liquidity
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [7d70a1c]
+- Updated dependencies [7d70a1c]
+- Updated dependencies [7d70a1c]
+  - @xchainjs/xchain-midgard@0.5.8
+  - @xchainjs/xchain-thornode@0.3.17
+  - @xchainjs/xchain-midgard-query@0.1.20
+  - @xchainjs/xchain-thorchain-query@0.7.17
+  - @xchainjs/xchain-thorchain-amm@1.1.18
+
+## 1.0.18
+
+### Patch Changes
+
+- Updated dependencies [7df3870]
+- Updated dependencies [7df3870]
+- Updated dependencies [7df3870]
+- Updated dependencies [7df3870]
+- Updated dependencies [7df3870]
+  - @xchainjs/xchain-thorchain-amm@1.1.17
+  - @xchainjs/xchain-thorchain-query@0.7.16
+
 ## 1.0.17
 
 ### Patch Changes